feat: make redirection work after signup
