+++ lib/sso.rb
@@ -1,0 +1,95 @@
+module DebtCollective
+  class SSO
+    COOKIE_DOMAIN = SiteSetting.sso_cookie_domain
+    COOKIE_NAME = SiteSetting.sso_cookie_name
+    JWT_SECRET = SiteSetting.sso_jwt_secret
+
+    def initialize(user, cookies = {})
+      @user = user
+      @cookies = cookies
+    end
+
+    def generate_jwt
+      hmac_secret = JWT_SECRET
+      jwt_alg = "HS256"
+
+      token = JWT.encode(jwt_payload, hmac_secret, jwt_alg)
+    end
+
+    def set_jwt_cookie
+      domain = ENV["JWT_COOKIE_DOMAIN"] || ".lvh.me"
+      secure = Rails.env.production?
+
+      @cookies[COOKIE_NAME] = {
+        domain: COOKIE_DOMAIN,
+        expires: SiteSetting.maximum_session_age.hours.from_now,
+        httponly: true,
+        secure: SiteSetting.force_https,
+        value: generate_jwt,
+      }
+    end
+
+    def remove_jwt_cookie
+      @cookies.delete(COOKIE_NAME, domain: COOKIE_DOMAIN)
+    end
+
+    private
+
+    def user_avatar_url
+      avatar_url = @user.small_avatar_url
+
+      if @user.uploaded_avatar.present?
+        base_url = Discourse.store.external? ? "#{Discourse.store.absolute_base_url}/" : Discourse.base_url
+        avatar_url = "#{base_url}#{Discourse.store.get_path_for_upload(@user.uploaded_avatar)}"
+      end
+
+      avatar_url
+    end
+
+    def user_profile_background_url
+      if @user.user_profile.profile_background_upload.present?
+        profile_background_url = UrlHelper.absolute(upload_cdn_path(
+          @user.user_profile.profile_background_upload.url
+        ))
+      end
+    end
+
+    def user_card_background_url
+      if @user.user_profile.card_background_upload.present?
+        card_background_url = UrlHelper.absolute(upload_cdn_path(
+          @user.user_profile.card_background_upload.url
+        ))
+      end
+    end
+
+    def user_custom_fields
+      custom_fields = {
+        state: @user.custom_fields.fetch("user_field_1", "").to_s,
+        zip_code: @user.custom_fields.fetch("user_field_2", "").to_s,
+        phone_number: @user.custom_fields.fetch("user_field_3", "").to_s
+      }
+    end
+
+    def jwt_payload
+      groups = @user.groups.collect(&:name)
+
+      payload = {
+        active: @user.active,
+        admin: @user.admin?,
+        avatar_url: user_avatar_url,
+        card_background_url: user_card_background_url,
+        created_at: @user.created_at,
+        custom_fields: user_custom_fields,
+        email: @user.email,
+        external_id: @user.id,
+        groups: groups,
+        last_seen_at: @user.last_seen_at,
+        moderator: @user.moderator?,
+        name: @user.name,
+        profile_background_url: user_profile_background_url,
+        updated_at: @user.updated_at,
+        username: @user.username,
+      }
+    end
+  end
+end
